changed components folder layout to split components between feature
