+++ client/src/components/assignments/AssignmentsDisplay.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from 'react'
 import '../styles/assignments.css'
-import assignmentsApi from '../api/assignmentsApi'
-import useAuth from '../hooks/useAuth'
-import useClassroom from '../hooks/useClassroom'
+import assignmentsApi from '../../api/assignmentsApi'
+import useAuth from '../../hooks/useAuth'
+import useClassroom from '../../hooks/useClassroom'
 import CreateAssignmentModal from './CreateAssignmentModal'
-import { UserRole } from '../utils/UserRole'
+import { UserRole } from '../../utils/UserRole'
 import StudentAssignments from './StudentAssignments'
 import ProfessorAssignments from './ProfessorAssignments'
