+++ client/src/components/files/FileItemModal.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faArrowDown } from '@fortawesome/free-solid-svg-icons'
-import { FileType } from '../utils/FileType'
-import fileSystemApi from '../api/fileSystemApi'
-import useAuth from '../hooks/useAuth'
-import useClassroom from '../hooks/useClassroom'
+import { FileType } from '../../utils/FileType'
+import fileSystemApi from '../../api/fileSystemApi'
+import useAuth from '../../hooks/useAuth'
+import useClassroom from '../../hooks/useClassroom'
 import '../styles/modal.css'
 
 interface addFileItemProps {
