+++ client/src/components/files/FilesDisplay.tsx
@@ -2,9 +2,9 @@ import { useEffect, useState } from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faArrowLeftLong } from '@fortawesome/free-solid-svg-icons'
 import FileItemModal from './FileItemModal'
-import useClassroom from '../hooks/useClassroom'
-import fileSystemApi from '../api/fileSystemApi'
-import useAuth from '../hooks/useAuth'
+import useClassroom from '../../hooks/useClassroom'
+import fileSystemApi from '../../api/fileSystemApi'
+import useAuth from '../../hooks/useAuth'
 import FileSystemItem from './FileSystemItem'
 import '../styles/file-system.css'
