+++ client/src/components/submissions/StudentSubmission.tsx
@@ -1,13 +1,13 @@
 import { useEffect, useState } from "react"
-import useClassroom from "../hooks/useClassroom"
+import useClassroom from "../../hooks/useClassroom"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faArrowUpFromBracket, faClock, faCircleCheck } from "@fortawesome/free-solid-svg-icons"
 import '../styles/submission.css'
 import SubmissionFileItem from "./SubmissionFileItem"
-import submissionApi from "../api/submissionApi"
-import useAuth from "../hooks/useAuth"
-import assignmentsApi from "../api/assignmentsApi"
-import { UserRole } from "../utils/UserRole"
+import submissionApi from "../../api/submissionApi"
+import useAuth from "../../hooks/useAuth"
+import assignmentsApi from "../../api/assignmentsApi"
+import { UserRole } from "../../utils/UserRole"
 
 const StudentSubmission = () => {
   const [dueDate, setDueDate] = useState<any>()
